Placing start and finish for max distance.
To make it difficult, we now compute the longest simple path through the maze
and position start and finish at its opposite ends.

Seems that this prevents the occasional very easy maze, but still makes it
pretty doable for sizes 3 through 5. With 6 it starts to get tricky, but still
won't keep you busy for hours.