Formatted HTML, now hosting GitHub ribbon
If we're not going to jeopardize player privacy by including external
resources (and thereby taking the opportunity for nice personalized Facebook
messages away), we might as well go all the way. The real reason behind this
policy of course being that I developed this game during long train rides
without Internet and was bitten by external resources not loading.